# ENH: stats.mode: vectorize implementation