@j-bowhay 2025-03-13
Given it is only two lines, it does not seem too bad to keep a separate 1D path