@mdhaber 2025-03-13
Done! Confirmed locally that performance for 1-D arrays is preserved.